requirements.txt
flask
requests
lxml
reportlab
python-docx
